More gcc fixes (missing consts, etc.)

+++ src/Gameplay.cpp
@@ -2429,7 +2429,7 @@ float get_projectile_range(point const &pos, vector3d vcf, float dist, float ran
 
 	if (coll) {
 		coll_obj &cobj(coll_objects.get_cobj(cindex));
-		if (cobj.cp.coll_func) {cobj.cp.coll_func(cobj.cp.cf_index, 0, zero_vector, pos, 0.0, (check_splash ? PROJC : LASER));} // apply collision function
+		if (cobj.cp.coll_func) {cobj.cp.coll_func(cobj.cp.cf_index, 0, zero_vector, pos, 0.0, (check_splash ? (int)PROJC : (int)LASER));} // apply collision function
 		range = p2p_dist(coll_pos, pos);
 	}
 	if (splash_val > 0.0 && is_underwater(pos1)) {gen_line_of_bubbles(pos1, (pos + vcf*range));}
